Internals & Design

File Interface

The JLDFile object mimics the API of Base.Dict as much as it can. In particular, keys, length, haskey, isempty, get, get! should work as expected.

B-tree Functions

JLD2.BTrees.write_v2btree_chunked_dataset — Function

write_v2btree_chunked_dataset(f::JLDFile, chunk_records, is_filtered::Bool,
                              node_size::UInt32=UInt32(2048),
                              split_percent::UInt8=UInt8(100),
                              merge_percent::UInt8=UInt8(40)) -> header_offset

Write a complete V2 B-tree chunk index structure (depth=0 only).

JLD2.BTrees.read_symbol_table_node — Function

read_symbol_table_node(f, offset)

Read a symbol table node which contains links to objects in old-style groups. Symbol table nodes are the leaf nodes of group B-trees.

JLD2.BTrees.read_records_in_node — Function

read_records_in_node(f, offset, num_records, depth, bh, record_reader)

Recursively read all records from a V2 B-tree node and its descendants. Returns a flat vector of all records in the subtree. Uses a callback function record_reader(io, type) for reading records.

JLD2.BTrees.read_v2btree_header — Function

read_v2btree_header(f, offset)

Read a V2 B-tree header from the file. Returns a V2BTreeHeader structure containing tree metadata. Checksum is validated but not returned as part of the struct.

Chunked Array Internals

JLD2.Chunking.read_chunked_array — Function

Read chunked dataset into preallocated array, dispatching by layout version and indexing type.

JLD2.TYPE_AS_DATA — Constant

TYPE_AS_DATA::ScopedValue{Bool}

Signal to the jlconvert function that the type being read will be treated as either data or a type. This is needed to allow a custom typemap function to return an Upgrade object when the type is going to be used for reconstructing an instance. (as a type)

JLD2.AbstractLink — Type

AbstractLink

Abstract base type for all link types in JLD2. Links represent connections between names in groups and objects or paths in the HDF5 file structure.

The concrete subtypes are:

HardLink: Direct pointer to an object header (traditional JLD2 behavior)

SoftLink: Path string resolved at access time within the same file

ExternalLink: Reference to an object in a different HDF5 file

See the HDF5 specification section 3.2.J for complete link message format details.

JLD2.CommittedDatatype — Type

CommittedDatatype <: H5Datatype

Reference to a shared datatype message (stored elsewhere in a file). These are stored in the _types group and indexed.

JLD2.CustomSerialization — Type

CustomSerialization{T,S}

On-disk representation for data that is written as if it were of Julia type T, but is read as type S.

JLD2.ExternalLink — Type

ExternalLink <: AbstractLink

Represents an external link that references an object in a different HDF5 file. This enables cross-file data dependencies and modular file organization.

Fields

file_path::String: Path to the external HDF5 file (relative or absolute)

object_path::String: Path within the external file to the target object

Example

# Link to /data/temperature in external_file.h5
link = ExternalLink("./external_file.h5", "/data/temperature")

JLD2.GlobalHeap — Type

GlobalHeap

Represents an HDF5 global heap structure.

JLD2.Group — Type

Group(file::JLDFile, name::String)
Group(file::Group, name::String)

Construct a Group in file with name name. Groups are JLD2s equivalent of folders and may be nested, so file itself may alread be a Group or a JLDFile file handle.

Example usage

jldopen("example.jld2", "w") do f
    g = Group(f, "subgroup")
    g["data"] = 42
end

jldopen("example.jld2") do f
    g = f["subgroup"]
    f["subgroup/data"] == g["data"]
end

Keyword arguments:

est_num_entries::Int = 4

est_link_name_len::Int = 8

Determine how much (additional) empty space should be allocated for the group description. (list of entries) This can be useful for performance when one expects to append many additional datasets after first writing the file.

JLD2.Group — Method

Group(f::JLDFile, name::AbstractString)

Construct an empty group named name at the top level of JLDFile f.

JLD2.Group — Method

Group(g::Group, name::AbstractString)

Construct a group named name as a child of group g.

JLD2.H5Datatype — Type

abstract type H5Datatype

Supertype of all HDF5 datatypes.

JLD2.HardLink — Type

HardLink <: AbstractLink

Represents a hard link that points directly to an object header within the same file.

JLD2.HeaderMessage — Type

HeaderMessage

Helper struct to read and write the first part of a header message.

JLD2.HeaderMessageIterator — Type

mutable struct HeaderMessageIterator{IO}
HeaderMessageIterator(f::JLDFile, offset::RelOffset)

Implements an iterator over header messages.

JLD2.Hmessage — Type

Hmessage{IO}

Representation of a Header Message in memory. Provides getproperty access to the fields of the message. Can also be used to construct and write custom messages.

JLD2.IndirectPointer — Type

IndirectPointer

When writing data, we may need to enlarge the memory mapping, which would invalidate any memory addresses arising from the old mmap pointer. IndirectPointer holds an offset relative to the MemoryBackedIO. It defers computing a memory address until converted to a Ptr{T}, so the memory mapping can be enlarged and addresses will remain valid.

JLD2.InlineUnionEl — Type

InlineUnionEl{T1,T2}(mask::UInt8, t1::T1, t2::T2)

Custom serialization struct for two member isbits union fields e.g. in other structs or arrays. To indicate that t1 is relevant the mask takes the value UInt8(0) and for t2 the mask takes the value UInt8(255).

JLD2.JLDFile — Type

JLDFile{T<:IO}

JLD file object.

JLD2.JLDWriteSession — Type

JLDWriteSession{T}

A JLDWriteSession keeps track of references to serialized objects. If T is a Dict, h5offset maps an object ID (returned by calling objectid) to th RelOffset of the written dataset. If it is Union{}, then references are not tracked, and objects referenced multiple times are written multiple times.

JLD2.MemoryBackedIO — Type

MemoryBackedIO <: IO

Abstract type for IO objects that are backed by memory in such a way that one can use pointer based unsafe_load and unsafe_store! operations after ensuring that there is enough memory allocated.

It needs to provide:

getproperty(io, :curptr) to get the current pointer

ensureroom(io, nb) to ensure that there are at least nb bytes available

position(io) to get the current (zero-based) position

seek(io, pos) to set the current position (zero-based)

JLD2.Message — Type

Message{IO}

Representation of a Message in memory. Provides getproperty access

JLD2.ReadRepresentation — Type

ReadRepresentation{T,ODR}

A type encoding both the Julia type T and the on-disk (HDF5) representation ODR.

JLD2.RelOffset — Type

RelOffset

Represents an HDF5 relative offset. This differs from a file offset (used elsewhere) in that it is relative to the superblock base address. fileoffset and h5offset convert between RelOffsets and file offsets.

JLD2.SharedDatatype — Type

SharedDatatype <: H5Datatype

Reference to a shared datatype message (stored elsewhere in a file).

JLD2.SoftLink — Type

SoftLink <: AbstractLink

Represents a soft link that contains a path string resolved at access time. The path can be absolute (starting with '/') or relative to the containing group.

Fields

path::String: The path to resolve when the link is accessed

JLD2.Upgrade — Type

Upgrade(T)

Specify an upgrade path for serialized structs using the typemap keyword argument and rconvert.

JLD2.attributes — Method

attributes(dset::Dataset; plain::Bool=false)

Return the attributes of a dataset as an OrderedDict. If plain is set to true then the values are returned as stored in the dataset object.

JLD2.behead — Method

behead(T)

Given a UnionAll type, recursively eliminates the where clauses

JLD2.bufferpos — Method

bufferpos(io::Union{BufferedReader, BufferedWriter})

Get the current position in the buffer.

JLD2.construct_array — Method

construct_array(io::IO, eltype, ndims::Int)

Construct array by reading ndims dimensions from io. Assumes io has already been seeked to the correct position.

JLD2.constructrr — Function

constructrr(f::JLDFile, T::DataType, dt::CompoundType, attrs::Vector{ReadAttribute},
            hard_failure::Bool=false)

Constructs a ReadRepresentation for a given type. This is the generic method for all types not specially handled below.

If hard_failure is true, then throw a TypeMappingException instead of attempting reconstruction. This helps in cases where we can't know if reconstructed parametric types will have a matching memory layout without first inspecting the memory layout.

JLD2.create_dataset — Method

create_dataset(parent, path, datatype, dataspace; kwargs...)

Arguments: - parent::Union{JLDfile, Group}: Containing group of new dataset - path: Path to new dataset relative to parent. If path is nothing, the dataset is unnamed. - datatype: Datatype of new dataset (element type in case of arrays) - dataspace: Dimensions or Dataspace of new dataset

Keyword arguments: - layout: DataLayout of new dataset - filters: FilterPipeline for describing the compression pipeline

JLD2.create_external_link! — Method

create_external_link!(group::Group, link_name::String, file_path::String, object_path::String)

Create an external link in the group that points to an object in another HDF5/JLD2 file.

Arguments

group::Group: The group to create the link in

link_name::String: Name of the link within the group

file_path::String: Path to the external HDF5/JLD2 file

object_path::String: Path to the object within the external file

Example

using JLD2

# Open a file and create an external link
jldopen("main.jld2", "w") do file
    create_external_link!(file, "external_data", "data.jld2", "/dataset1")
end

Note: External file paths are validated to prevent directory traversal attacks.

JLD2.create_soft_link! — Method

create_soft_link!(group::Group, link_name::String, target_path::String)

Create a soft link (symbolic link) within the same HDF5/JLD2 file.

Arguments

group::Group: The group to create the link in

link_name::String: Name of the link within the group

target_path::String: Path to the target object within the same file

Example

using JLD2

jldopen("test.jld2", "w") do file
    file["original_data"] = [1, 2, 3, 4, 5]
    create_soft_link!(file, "link_to_data", "/original_data")
end

JLD2.default_typemap — Method

default_typemap(f::JLDFile, typepath::String, params)

Default type mapping function used by JLD2 to resolve data types read from files.

Arguments:

f::JLDFile: The JLD file being read.

typepath::String: The path to the type as a string, e.g. "Main.MyModule.MyType".

params: A list of type parameters for the type (may be empty).

JLD2.fileoffset — Method

fileoffset(f::JLDFile, x::RelOffset)

Converts an offset x relative to the superblock of file f to an absolute offset.

JLD2.find_type — Method

find_type(typepath::String)

Finds a type in the loaded modules by its path as a string. The type path should be a dot-separated string, e.g. "Main.MyModule.MyType". If the type is found, it returns the corresponding DataType or UnionAll. If the type is not found, it returns nothing.

JLD2.flag2uint — Method

flag2uint(flag::UInt8)

I Map the lowest to bits of flag to a UInt type, mapping 0 to UInt8, 1 to UInt16, 2 to UInt32, and 3 to UInt64.

JLD2.get_dataset — Method

get_dataset(parent::Union{JLDFile, Group}, name::String)

Get a stored dataset from a file by name or path as a Dataset object. This may be useful for inspecting the metadata incl. types of a dataset.

JLD2.h5offset — Method

h5offset(f::JLDFile, x::Integer)

Converts an absolute file offset x to an offset relative to the superblock of file f.

JLD2.ismmappable — Method

ismmappable(dset::Dataset)

Check if a dataset can be memory-mapped. This can be useful for large arrays and for editing written arrays.

An Array dataset may be mmapped if: - JLD2.samelayout(T) == true: The element type is isbits and has a size that either 1, 2, 4, or a multiple of 8 bytes. - Uncompressed: Compressed arrays cannot be memory-mapped - Uses a contiguous layout: This is true for all array datasets written by JLD2 with version ≥ v0.4.52 - Windows: The file must be opened in read-only mode. This is a limitation of Mmap on Windows.

JLD2.isset — Method

isset(flag, bit)

Return true if the bit-th bit of flag is set. (starting from 0)

JLD2.jld_finalizer — Method

jld_finalizer(f::JLDFile)

When a JLDFile is finalized, it is possible that the MmapIO has been munmapped, since Julia does not guarantee finalizer order. This means that the underlying file may be closed before we get a chance to write to it.

JLD2.jldopen — Function

jldopen(file, mode::AbstractString; iotype=MmapIO, compress=false, typemap=JLD2.default_typemap)

Opens a JLD2 file at path file. Alternatively file may be a suitable IO object.

Options for mode:

"r": Open for reading only, failing if no file exists

"r+": Open for reading and writing, failing if no file exists

"w"/"w+": Open for reading and writing, overwriting the file if it already exists

"a"/"a+": Open for reading and writing, creating a new file if none exists, but preserving the existing file if one is present

JLD2.jldsave — Function

jldsave(filename; kwargs...)
jldsave(filename, compress; kwargs...)
jldsave(filename, compress, iotype; kwargs...)

Creates a JLD2 file at filename and stores the variables given as keyword arguments.

Examples

jldsave("example.jld2"; a=1, b=2, c)

is equivalent to

jldopen("example.jld2", "w") do f
    f["a"] = 1
    f["b"] = 2
    f["c"] = c
end

To choose the io type IOStream instead of the default MmapIO use jldsave(fn, IOStream; kwargs...).

JLD2.jlwrite — Method

jlwrite(io::IO, x::Tuple)

Attempt to write a tuple to io by writing each element of the tuple in order.

JLD2.links_size — Method

links_size(pairs)

Returns the size of several link messages. pairs is an iterator of String => AbstractLink pairs.

JLD2.load_attributes — Function

load_attributes(f::JLDFile, name::AbstractString)
load_attributes(g::Group, name::AbstractString)
load_attributes(g::Group)
load_attributes(f::JLDFile, offset::RelOffset)

Return a list of attributes attached to the dataset or group.

JLD2.load_data_or_dict — Method

load_data_or_dict(g::Union{JLDFile,Group}, varname::AbstractString)

Return the value of key varname but if it represents a Group load the group as a nested dictionary.

JLD2.load_datatypes — Method

load_datatypes(f::JLDFile)

Populate f.datatypes and f.jlh5types with all of the committed datatypes from a file. We need to do this before writing to make sure we reuse written datatypes.

JLD2.load_object — Method

load_object(filename)

Returns the only available object from the JLD2 file filename (The stored object name is inconsequential). If the file contains more than one or no objects, the function throws an ArgumentError.

For loading more than one object, use @load macro, jldopen or the FileIO API.

Example

To load the only object from the JLD2 file example.jld2:

hello = "world"
save_object("example.jld2", hello)
hello_loaded = load_object("example.jld2")

JLD2.loadnesteddict — Method

loadnesteddict(g::Union{JLDFile, Group})

Return a dictionary with all data contained in group or file. Nested groups are loaded as nested dictionaries.

JLD2.lookup_link — Method

lookup_link(g::Group, name::AbstractString) -> Union{AbstractLink, Nothing}

Lookup a link in a group by name. Returns the AbstractLink object if found, or nothing if not present. This function supports all link types (hard, soft, external).

JLD2.message_size_for_link — Method

message_size_for_link(name::String, link::AbstractLink) -> Int

Calculate the size of a link message for the given link type.

JLD2.parse_link_message — Method

parse_link_message(wmsg::HmWrap{HmLinkMessage}) -> AbstractLink

Parse a link message from the HDF5 format and return the appropriate AbstractLink subtype. Handles hard links (type 0), soft links (type 1), and external links (type 64).

JLD2.pathize — Method

pathize(g::Group, name::AbstractString, create::Bool) -> Tuple{Group,String}

Converts a path to a group and name object. If create is true, any intermediate groups will be created, and the dataset name will be checked for uniqueness with existing names.

JLD2.prewrite — Method

prewrite(f::JLDFile)

Check that a JLD file is actually writable, and throw an error if not. Sets the written flag on the file.

JLD2.print_header_messages — Method

print_header_messages(f::JLDFile, name::AbstractString)
print_header_messages(g::Group, name::AbstractString)
print_header_messages(f::JLDFile, offset::RelOffset)

Prints the header messages of a group or dataset in a file.

JLD2.printtoc — Method

printtoc([io::IO,] f::JLDFile [; numlines])

Prints an overview of the contents of f to the IO.

Use the optional numlines parameter to restrict the amount of items listed.

JLD2.read_array! — Function

read_array!(v::Array, f::JLDFile, rr)

Fill the array v with the contents of JLDFile f at the current position, assuming a ReadRepresentation rr.

JLD2.read_attr_data — Method

read_attr_data(f::JLDFile, attr::ReadAttribute, expected_datatype::H5Datatype,
               rr::ReadRepresentation)

jlread data from an attribute, assuming a specific HDF5 datatype and ReadRepresentation. If the HDF5 datatype does not match, throws an UnsupportedFeatureException. This allows better type stability while simultaneously validating the data.

JLD2.read_attr_data — Method

read_attr_data(f::JLDFile, attr::ReadAttribute)

jlread data from an attribute.

JLD2.read_compressed_array! — Function

read_compressed_array!(v::Array, f::JLDFile, rr, data_length::Int, filter, filter_mask=0)

Fill the array v with the compressed contents of JLDFile f at the current position, assuming a ReadRepresentation rr and that the compressed data has length data_length.

The optional filter_mask parameter indicates which filters were skipped during compression (bit n set = filter n was skipped). When filter_mask is non-zero, only filters not masked will be applied during decompression. A filter_mask of 0 (default) means all filters are applied.

JLD2.read_data — Function

read_data(f::JLDFile, dataspace::ReadDataspace, datatype_class::UInt8,
          datatype_offset::Int64, data_offset::Int64[, filters::FilterPipeline,
          header_offset::RelOffset, attributes::Vector{ReadAttribute}])

Read data from a file. If datatype_class is typemax(UInt8), the datatype is assumed to be committed, and datatype_offset points to the offset of the committed datatype's header. Otherwise, datatype_offset points to the offset of the datatype attribute.

JLD2.read_dataset — Method

read_dataset(dset::Dataset)

Read the data referenced by a dataset.

JLD2.read_fractal_heap_group — Method

read_btree(f, offset_hh, offset_bh)

Read a complete V2 B-tree for indexed groups with fractal heap storage. Returns a vector of tuples: (link_name::String, target::RelOffset)

JLD2.read_scalar — Function

read_scalar(f::JLDFile, rr, header_offset::RelOffset)

Read raw data representing a scalar with read representation rr from the current position of JLDFile f. header_offset is the RelOffset of the object header, used to resolve cycles.

JLD2.read_size — Method

read_size(io::IO, flags::UInt8)

Loads a variable-length size according to flags

Expects that the first two bits of flags mean:

0: The size of the Length of Link Name field is 1 byte.

1: The size of the Length of Link Name field is 2 bytes.

2: The size of the Length of Link Name field is 4 bytes.

3: The size of the Length of Link Name field is 8 bytes.

Returns the size as an Int.

JLD2.readas — Method

readas(::Type)::Type

Experimental feature: JLD2.readas can be overloaded to override which type a saved type is read as, and is used together with custom serialization using JLD2.writeas.

The typical case is custom serialization of parametric types, where not all type parameters are available during reading. Consider the following example for an anonymous function fun inside a Foo

struct Foo{F<:Function}
    fun::F
end
struct FooSerialization
    fun
end
JLD2.writeas(::Type{<:Foo}) = FooSerialization
Base.convert(::Type{<:FooSerialization}, f::Foo) = FooSerialization(f.fun)

JLD2.readas(::Type{<:FooSerialization}) = Foo
struct UndefinedFunction <:Function
    fun
end
(f::UndefinedFunction)(args...; kwargs...) = error("The function $(f.fun) is not defined")
function Base.convert(::Type{<:Foo}, f::FooSerialization)
    isa(f.fun, Function) && return Foo(f.fun)
    return Foo(UndefinedFunction(f.fun))
end

If we include these definitions, call jldsave("foo.jld2"; foo=Foo(x->x^2)), restart julia, include the definitions again, and call foo = jldopen("foo.jld2") do io; io["foo"]; end, we get foo::Foo{UndefinedFunction} and foo::FooSerialization with and without defining the JLD2.readas above, respectively.

JLD2.readmmap — Method

readmmap(dset::Dataset)

Memory-map a dataset. This can be useful for large arrays and for editing written arrays. See ismmappable for requirements.

JLD2.save_group — Method

save_group(g::Group) -> RelOffset

Stores a group to a file, updating it if it has already been saved. Returns UNDEFINED_ADDRESS if the group was already stored, or the offset of the new group otherwise.

JLD2.save_object — Method

save_object(filename, x)

Stores an object x in a new JLD2 file at filename. If a file exists at this path, it will be overwritten.

Since the JLD2 format requires that all objects have a name, the object will be stored as single_stored_object. If you want to store more than one object, use @save macro, jldopen or the FileIO API.

Example

To save the string hello to the JLD2 file example.jld2:

hello = "world"
save_object("example.jld2", hello)

JLD2.shorttypestring — Method

shorttypestring(::Type{ <:UnknownType})

Convert an UnknownType to a corresponding string. This is only used to create names for reconstructed types.

See also typestring.

JLD2.size_flag — Method

size_flag(sz::Integer)::UInt8

Return the flag that represents the smallest integer type that can represent sz. 0 -> UInt8, 1 -> UInt16, 2 -> UInt32, 3 -> UInt64

JLD2.size_size — Method

size_size(sz::Integer)

Return the number of bytes required to represent sz as an unsigned integer that actually exists. (e.g. UInt8, UInt16, UInt32, UInt64)

JLD2.size_size2 — Method

size_size2(sz::Integer)

Return the number of bytes required to represent sz as an unsigned integer. Note: this does not check if the integer is a valid julia integer.

JLD2.skip_to_aligned! — Function

skip_to_aligned!(io, rel=0)

Skip to nearest position aligned to a multiple of 8 bytes relative to rel.

JLD2.symbol_length — Method

symbol_length(x::Symbol)

Returns the length of the string represented by x.

JLD2.typestring — Method

typestring(::Type{ <:UnknownType})

Convert an UnknownType to a corresponding string. This is only used for warning during reconstruction errors.

See also shorttypestring.

JLD2.uintofsize — Method

uintofsize(sz::Integer)

Return the UInt type that has sz bytes.

JLD2.write_dataset — Function

write_dataset(dataset::Dataset, data)

Write data to file using metadata prepared in the dataset.

JLD2.write_link_message — Method

write_link_message(io, name::String, link::AbstractLink)

Write a link message for the given link type to the I/O stream.

JLD2.write_size — Method

write_size(io::IO, sz::Integer)

Write the mininum number of bytes required to represent sz as (valid) unsigned integer.

JLD2.write_zerobytes — Method

write_zerobytes(io, n)

Write n zero bytes to io.

JLD2.@load — Macro

@load filename var1 [var2 ...]

Load one or more variables var1,... from JLD2 file filename into the current scope and return a vector of the loaded variable names.

For interactive use, the form @load "somefile.jld2" will load all variables from "somefile.jld2" into the current scope. This form only supports literal file names and should be avoided in more permanent code so that it's clear where the variables come from.

Example

To load the variables hello and foo from the file example.jld2, use

@load "example.jld2" hello foo

JLD2.@pseudostruct — Macro

@pseudostruct name begin ... end

The @pseudostruct macro is used to define constructor, size computation, show, and and optimized getproperty function for Messages. The allowed syntax elements are:

@skip(n): Mark n bytes as empty.

JLD2.@save — Macro

@save filename var1 [var2 ...]
@save filename {compress=true} var1 name2=var2

Write one or more variables var1,... from the current scope to a JLD2 file filename.

For interactive use you can save all variables in the current module's global scope using @save filename. More permanent code should prefer the explicit form to avoid saving unwanted variables.

Example

To save the string hello and array xs to the JLD2 file example.jld2:

hello = "world"
xs = [1,2,3]
@save "example.jld2" hello xs

For passing options to the saving command use {}

@save "example.jld2" {compress=true} hello xs

For saving variables under a different name use regular assignment syntax

@save "example.jld2" greeting=hello xarray = xs